изменение типов в ги при посылке инвайта проверяем нахождение в ги отключаем вывод имени чара при ошибках ги для предотвращения крашей клиента

+++ src/server/game/Entities/Creature/Creature.h
@@ -421,7 +421,7 @@ struct VendorItemData
         return m_items[slot];
     }
     bool Empty() const { return m_items.empty(); }
-    uint8 GetItemCount() const { return m_items.size(); }
+    uint32 GetItemCount() const { return m_items.size(); }
     void AddItem(uint32 item, int32 maxcount, uint32 ptime, uint32 ExtendedCost, uint8 type)
     {
         m_items.push_back(new VendorItem(item, maxcount, ptime, ExtendedCost, type));
