+++ src/server/game/Handlers/ItemHandler.cpp
@@ -747,7 +747,7 @@ void WorldSession::SendListInventory(uint64 vendorGuid)
         vendor->StopMoving();
 
     VendorItemData const* vendorItems = vendor->GetVendorItems();
-    uint8 rawItemCount = vendorItems ? vendorItems->GetItemCount() : 0;
+    uint32 rawItemCount = vendorItems ? vendorItems->GetItemCount() : 0;
 
     //if (rawItemCount > 300),
     //    rawItemCount = 300; // client cap but uint8 max value is 255
@@ -757,9 +757,9 @@ void WorldSession::SendListInventory(uint64 vendorGuid)
     enablers.reserve(2 * rawItemCount);
 
     const float discountMod = _player->GetReputationPriceDiscount(vendor);
-    uint8 count = 0;
-    uint8 realCount = 0;
-    for (uint8 slot = 0; slot < rawItemCount; ++slot)
+    uint32 count = 0;
+    uint32 realCount = 0;
+    for (uint32 slot = 0; slot < rawItemCount; ++slot)
     {
         count++;
